Unreviewed, fix test failure and add additional tests after r248753
* inspector/unit-tests/url-utilities.html:
* inspector/unit-tests/url-utilities-expected.txt:
Add additional tests of `WI.displayNameForURL` with paths, query strings, and fragments.

* inspector/css/add-rule.html:
* inspector/css/add-rule-expected.txt:
The `doNotCreateIfMissing` parameter was removed from `preferredInspectorStyleSheetForFrame`.
All the test is trying to do is check that there's not an existing Inspector Style Sheet, so
instead just check that the list of Inspector Style Sheets is empty.

Identifier: 214549@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248774 268f45cc-cd09-0410-ab3c-d52691b4dbfc