Layout tests that call resizeTo() crash when run on iOS with IOSurface support enabled https://bugs.webkit.org/show_bug.cgi?id=200866 <rdar://problem/50254021>
Reviewed by Simon Fraser.

Source/WebKit:

* UIProcess/ios/PageClientImplIOS.mm:
(WebKit::PageClientImpl::convertToDeviceSpace):
(WebKit::PageClientImpl::convertToUserSpace):
Implement convertTo{Device,User}Space in PageClientImplIOS.
We don't actually do any conversion. For our purposes, the window is
"device" space, and we never allow iOS WebKit clients to resize the window...
except for in tests! So just pass the rects straight through, instead of
returning an empty rect.

Tools:

* WebKitTestRunner/ios/PlatformWebViewIOS.mm:
(WTR::PlatformWebView::windowSnapshotImage):
Replace some logging with release assertions that we are snapshotting a reasonable
view and get a reasonable snapshot back. Failing to snapshot is a big deal, and
"silently" failing with just a log that will be ignored is not helpful.


Identifier: 214602@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248828 268f45cc-cd09-0410-ab3c-d52691b4dbfc