[LFC][TFC] Compute column min/max widths and table width. https://bugs.webkit.org/show_bug.cgi?id=200757 <rdar://problem/54333148>
Reviewed by Antti Koivisto.

For each column, determine a maximum and minimum column width from the cells that span only that column.
The minimum is that required by the cell with the largest minimum cell width (or the column 'width', whichever is larger).
The maximum is that required by the cell with the largest maximum cell width (or the column 'width', whichever is larger).

* layout/tableformatting/TableFormattingContext.cpp:
(WebCore::Layout::TableFormattingContext::computePreferredWidthForColumns const):
* layout/tableformatting/TableGrid.cpp:
(WebCore::Layout::TableGrid::appendCell):
* layout/tableformatting/TableGrid.h:
(WebCore::Layout::TableGrid::columns):
(WebCore::Layout::TableGrid::rows):


Identifier: 214608@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@248834 268f45cc-cd09-0410-ab3c-d52691b4dbfc