Have RenderLayer::calculateClipRects() use offsetFromAncestor() when possible https://bugs.webkit.org/show_bug.cgi?id=201066
Reviewed by Dean Jackson.

offsetFromAncestor() is a layer tree walk, so faster than localToContainerPoint(),
but we can't use it when there are transforms on the layer and intermediates up
to the ancestor.

canUseConvertToLayerCoords() was trying to answer the question about whether it's
OK to use offsetFromAncestor() (which calls convertToLayerCoords() internally), but
it has insufficient information to make a determination. Leave this issue alone, but
at least rename canUseConvertToLayerCoords().

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::updateLayerPositions):
(WebCore::RenderLayer::calculateClipRects const):
* rendering/RenderLayer.h:


Identifier: 214814@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249091 268f45cc-cd09-0410-ab3c-d52691b4dbfc