Optimize computation of AbsoluteClipRects clip rects https://bugs.webkit.org/show_bug.cgi?id=201148
Reviewed by Zalan Bujtas.

When adding layers to the compositing overlap map, we compute AbsoluteClipRects for every
layer which is expensive. This was more expensive than necessary because we converted them
to TemporaryClipRects when crossing painting boundaries, but AbsoluteClipRects don't
care about painting boundaries, so don't do this.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateClipRects const):


Identifier: 214841@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249120 268f45cc-cd09-0410-ab3c-d52691b4dbfc