webkitpresentationmodechanged is fired twice when exiting picture in picture https://bugs.webkit.org/show_bug.cgi?id=193765
Patch by Peng Liu <peng.liu6@apple.com> on 2019-08-27
Reviewed by Jer Noble.

Source/WebCore:

This patch removes the extra "webkitpresentationmodechanged" event when the browser switches from
picture-in-picture or fullscreen to inline.

The bug was introduced by the fix for bug
https://bugs.webkit.org/show_bug.cgi?id=181095
But now we are using modern media controls and the fix is not necessary.
Reverting that fix can fix the issue.

Also, this patch gets rid of the unnecessary try to call a JavaScript function which is not available
in the modern media controls.

Test: media/presentationmodechanged-fired-once.html

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::exitFullscreen):
(WebCore::HTMLMediaElement::updateMediaControlsAfterPresentationModeChange):

LayoutTests:

* media/presentationmodechanged-fired-once-expected.txt: Added.
* media/presentationmodechanged-fired-once.html: Added.

Identifier: 214862@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249141 268f45cc-cd09-0410-ab3c-d52691b4dbfc