fast/scrolling/ios/click-events-during-momentum-scroll-in-overflow-after-tap-on-body.html times out on iPad https://bugs.webkit.org/show_bug.cgi?id=201182 <rdar://problem/54562829>
Reviewed by Tim Horton.

This test verifies that if the user taps outside of a scrollable container that is decelerating, the page will
observe a click event. This test currently behaves as expected on iPhone, but on iPad, the tap which is
intended to be outside the scrollable area ends up inside it, so a click event is never dispatched. Instead of
hard-coding the tap location, address this test failure by targetting a separate element outside of the
scrollable area.

* fast/scrolling/ios/click-events-during-momentum-scroll-in-overflow-after-tap-on-body.html:


Identifier: 214868@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249152 268f45cc-cd09-0410-ab3c-d52691b4dbfc