Drawing an animated image to a canvas via drawImage should draw the first frame https://bugs.webkit.org/show_bug.cgi?id=74779 <rdar://problem/42282454>
Reviewed by Said Abou-Hallawa.

Source/WebCore:

After this patch, when an animated image is drawn into a canvas via
drawImage, the actual image that the canvas will receive will be
a new Image object with the contents of the first frame of the animation.

It also adds an internal setting to keep the previous behaviour
(draw the current frame playing) because there are several layout
tests that rely on canvas.drawImage() to check that animated images
play properly.

Test: fast/canvas/drawImage-animated-gif-draws-first-frame-and-no-reset-image.html

* html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::drawImage):
* page/Settings.yaml:
* testing/InternalSettings.cpp:
(WebCore::InternalSettings::Backup::Backup):
(WebCore::InternalSettings::Backup::restoreTo):
(WebCore::InternalSettings::setAnimatedImageDebugCanvasDrawingEnabled):
* testing/InternalSettings.h:
* testing/InternalSettings.idl:

LayoutTests:

Adds a test that checks that when drawing an animated image into a canvas
it draws the first frame and that the animation doesn't reset or pause.

Also enable the new internal setting setAnimatedImageDebugCanvasDrawingEnabled()
for the tests that rely on canvas.drawImage() drawing the current frame instead
of the first one.

* fast/canvas/drawImage-animated-gif-draws-first-frame-and-no-reset-image-expected.txt: Added.
* fast/canvas/drawImage-animated-gif-draws-first-frame-and-no-reset-image.html: Added.
The above test checks that when drawing an animated image to a canvas it draws the first frame
and that the playing image doesn't get reseted and finish playing.
* fast/canvas/resources/animated-red-green-blue-yellow-cyan-black-repeat-1.gif: Added.
* fast/images/animated-gif-restored-from-bfcache.html: Enable internals.settings.setAnimatedImageDebugCanvasDrawingEnabled().
* fast/images/animated-image-different-dest-size.html: Ditto.
* fast/images/animated-image-loop-count.html: Ditto.
* fast/images/animated-image-mp4.html: Ditto.
* fast/images/decode-animated-image.html: Ditto.
* fast/images/decode-render-animated-image.html: Ditto.
* fast/images/ordered-animated-image-frames.html: Ditto.
* fast/images/reset-image-animation.html: Ditto.
* fast/images/slower-animation-than-decoding-image.html: Ditto.
* fast/images/slower-decoding-than-animation-image.html: Ditto.


Identifier: 214878@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@249162 268f45cc-cd09-0410-ab3c-d52691b4dbfc